addProductToCart

+++ src/main/java/vn/baotran/laptopshop/domain/CartDetail.java
@@ -11,9 +11,9 @@ public class CartDetail {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    private Long quantity;
+    private Integer quantity;
 
-    private Long price;
+    private Double price;
 
     @ManyToOne
     @JoinColumn(name = "cart_id")
